[Android] Move extension to new file and remove unnecessary extensions as well
