+++ feature/news/src/main/kotlin/com/instantsystem/android/feature/news/domain/interactor/GetTopHeadlinesUseCase.kt
@@ -1,9 +1,9 @@
 package com.instantsystem.android.feature.news.domain.interactor
 
-import com.instantsystem.android.feature.news.data.entity.DefaultNewsResponse
 import com.instantsystem.android.feature.news.data.entity.TopHeadlinesRequest
 import com.instantsystem.android.feature.news.data.repository.NewsRepository
 import com.instantsystem.android.feature.news.domain.model.NewsArticle
+import com.instantsystem.android.feature.news.ext.toDomain
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
@@ -20,23 +20,10 @@ class GetTopHeadlinesUseCase(
                 TopHeadlinesRequest(
                     country = param.country
                 )
-            )?.toModel() ?: emptyList()
+            )?.articles?.toDomain() ?: emptyList()
         }
 }
 
 data class TopHeadlinesParam(
     val country: String
 )
-
-fun DefaultNewsResponse.toModel(): List<NewsArticle> = articles.map {
-    NewsArticle(
-        title = it.title.orEmpty(),
-        description = it.description.orEmpty(),
-        url = it.url.orEmpty(),
-        urlToImage = it.urlToImage.orEmpty(),
-        content = it.content?.take(200).orEmpty(),
-        author = it.author.orEmpty(),
-        publishedAt = it.publishedAt.orEmpty(),
-        source = it.source?.name.orEmpty()
-    )
-}
